+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 ### Bugfixes
 - Fixed water animations not being stopped upon exiting water.
+- Fixed crash that happens when a penguin has a boat to follow.